analytics_service/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
python-dotenv
requests
kubernetes
django-redis
redis
